pyproject.toml
[tool.poetry]
name = "telephonepi"
version = "0.1.0"
description = ""
authors = ["Niclas Olofsson <n@niclasolofsson.se>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.33.2"
google-assistant-grpc = "^0.3.0"
tenacity = ">=4.1.0,<5"
sounddevice = "^0.4.1"
click = "^7.1.2"
gpiozero = "^1.5.1"
"rpi.gpio" = "^0.7.0"

[tool.poetry.dev-dependencies]
google-auth-oauthlib = "^0.4.2"
google-assistant-sdk = "^0.6.0"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
